исправлена ошибка сохранения в бд в bottom_parser, добавлен шаблон таблицы отриц отзывов
