--- bank_bottom_proj/get_all_feedback.py
@@ -1,7 +1,0 @@
-from webapp_bottom import create_app
-from webapp_bottom.bottom_parser import page_fliper
-
-
-app = create_app()
-with app.app_context():
-    page_fliper()
