+++ bank_bottom_proj/webapp_bottom/__init__.py
@@ -2,11 +2,10 @@
 # source env_bottom/bin/activate
 # export FLASK_APP=webapp_bottom && flask db init
 
-from flask import Flask, render_template, flash, redirect, url_for
+from flask import Flask, render_template
 from webapp_bottom.model  import db, Feedback
 from flask_migrate import Migrate
 import datetime 
-
 
 def create_app():
     app = Flask(__name__)
@@ -20,7 +19,9 @@ def create_app():
         delta = datetime.timedelta(days=7)
         now_day = datetime.date.today()
         delta_days = now_day - delta
-        qr = db.session.query(Feedback.bank_name,db.func.count(Feedback.url_page)).filter(Feedback.response_date > delta_days).group_by(Feedback.bank_name).having(db.func.count(Feedback.url_page) > 10).order_by(db.func.count(Feedback.url_page).desc()).all()
+        date_obj = datetime.datetime.strptime(str(delta_days), '%Y-%m-%d')
+        formatted_date_str = date_obj.strftime('%d.%m.%Y')
+        qr = db.session.query(Feedback.bank_name,db.func.count(Feedback.url_page)).filter(Feedback.response_date >= formatted_date_str).group_by(Feedback.bank_name).having(db.func.count(Feedback.url_page) > 10).order_by(db.func.count(Feedback.url_page).desc()).all()
         return render_template('index.html', page_title=title, banks_list= qr)
     
     @app.route("/feedback")
